Add unittest for detect_faces function :koala:
