+++ .gitignore
@@ -1,3 +1,4 @@
 venv
 *.pyc
 *.jpg
+*.swp
